feat: implement domain events, event handlers, entity refactor and build fixes

+++ src/IzTek.Carbon.Footprint.Application/Features/DailyActivities/Queries/GetActivityCalendar/GetActivityCalendarQuery.cs
@@ -13,7 +13,7 @@ public record GetActivityCalendarQuery(
 
 public class CalendarResponse
 {
-    public int TotalScore { get; set; } // Seçilen yıla veya aya ait toplam skor
+    public double TotalScore { get; set; } // Seçilen yıla veya aya ait toplam skor
     public List<CalendarItemDto> Items { get; set; } = new();
 };
 
@@ -21,7 +21,7 @@ public class CalendarItemDto
 {
     public DateTime Date { get; set; }
 
-    public int Score { get; set; }
+    public double Score { get; set; }
 
     public bool HasDetails { get; set; }
 };
